feat: domain with and without scheme resolving
