+++ backend/internal/httpclient/httpClient.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"io"
 	"log/slog"
-	"net"
 	"net/http"
+	"net/url"
 	"time"
 
 	"github.com/lananolana/webpulse/backend/internal/config"
@@ -14,14 +14,15 @@ import (
 	"github.com/lananolana/webpulse/backend/pkg/logger/sl"
 )
 
+// Ptr returns a pointer to the given value.
 func Ptr[T any](v T) *T { return &v }
 
-// HttpClient represents HTTP client
+// HttpClient represents an HTTP client.
 type HttpClient struct {
 	client *http.Client
 }
 
-// New creates a new HTTP client
+// New creates a new HttpClient with the provided configuration.
 func New(cfg config.HTTPClient) *HttpClient {
 	return &HttpClient{
 		client: &http.Client{
@@ -31,19 +32,16 @@ func New(cfg config.HTTPClient) *HttpClient {
 	}
 }
 
-// GetServiceStats sends a request to the specified domain and returns service statistics
+// GetServiceStats sends a request to the specified domain and returns service statistics.
 func (c *HttpClient) GetServiceStats(ctx context.Context, domain string) dto.ServiceStatsResponse {
-	url := "https://" + domain
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
+	u, err := parseURL(domain)
 	if err != nil {
-		slog.Error("Failed to create request", sl.Err(err))
+		slog.Error("Failed to parse URL", sl.Err(err))
 		return newFailedCreateRequestResponse()
 	}
 
-	startRequest := time.Now()
-	resp, err := c.client.Do(req)
-	if err != nil || resp == nil {
-		slog.Error("Failed to perform request", sl.Err(err))
+	resp, startRequest, err := c.makeRequest(ctx, u)
+	if err != nil {
 		return newTargetUnavailableResponse()
 	}
 	defer closeResponseBody(resp.Body)
@@ -70,10 +68,10 @@ func (c *HttpClient) GetServiceStats(ctx context.Context, domain string) dto.Ser
 	corsInfo := getCORSInfo(resp)
 
 	startDNS := time.Now()
-	dnsRecords := getDNSDetails(domain)
+	dnsRecords := getDNSDetails(u.Host)
 	dnsResponseTimeMs := time.Since(startDNS).Milliseconds()
 
-	serverInfo, err := getServerInfo(domain, resp, dnsResponseTimeMs, dnsRecords)
+	serverInfo, err := getServerInfo(u.Host, resp, dnsResponseTimeMs, dnsRecords)
 	if err != nil {
 		slog.Error("Failed to get server info", sl.Err(err))
 		return newTargetUnavailableResponse()
@@ -91,7 +89,42 @@ func (c *HttpClient) GetServiceStats(ctx context.Context, domain string) dto.Ser
 	}
 }
 
-// Helper functions
+// makeRequest attempts to make an HTTP request to the given URL.
+func (c *HttpClient) makeRequest(ctx context.Context, u *url.URL) (*http.Response, time.Time, error) {
+	if u.Scheme != "" {
+		return c.doRequest(ctx, u)
+	}
+
+	schemes := []string{"https", "http"}
+	for _, scheme := range schemes {
+		u.Scheme = scheme
+		resp, startRequest, err := c.doRequest(ctx, u)
+		if err == nil {
+			return resp, startRequest, nil
+		}
+	}
+	slog.Error("Failed to perform request with both https and http schemes")
+	return nil, time.Time{}, http.ErrHandlerTimeout
+}
+
+// doRequest performs the HTTP request and returns the response.
+func (c *HttpClient) doRequest(ctx context.Context, u *url.URL) (*http.Response, time.Time, error) {
+	urlStr := u.String()
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, urlStr, nil)
+	if err != nil {
+		slog.Error("Failed to create request", sl.Err(err))
+		return nil, time.Time{}, err
+	}
+
+	startRequest := time.Now()
+	resp, err := c.client.Do(req)
+	if err != nil {
+		slog.Error("Failed to perform request", sl.Err(err))
+		return nil, startRequest, err
+	}
+
+	return resp, startRequest, nil
+}
 
 func newTargetUnavailableResponse() dto.ServiceStatsResponse {
 	return dto.ServiceStatsResponse{
@@ -111,101 +144,4 @@ func closeResponseBody(body io.ReadCloser) {
 		_, _ = io.Copy(io.Discard, body)
 		_ = body.Close()
 	}
-}
-
-func calculateTransferSpeed(bytesReceived int, durationMillis int64) int64 {
-	if durationMillis == 0 {
-		return 0
-	}
-	kilobits := float64(bytesReceived*8) / 1000.0
-	seconds := float64(durationMillis) / 1000.0
-	return int64(kilobits / seconds)
-}
-
-func createAvailability(resp *http.Response) dto.Availability {
-	return dto.Availability{
-		IsAvailable:    Ptr(true),
-		HTTPStatusCode: &resp.StatusCode,
-	}
-}
-
-func getSSLInfo(resp *http.Response) *dto.SSL {
-	if resp.TLS == nil || len(resp.TLS.PeerCertificates) == 0 {
-		return nil
-	}
-	valid := resp.TLS.HandshakeComplete
-	expiresAt := resp.TLS.PeerCertificates[0].NotAfter.Unix()
-	issuer := resp.TLS.PeerCertificates[0].Issuer.String()
-
-	return &dto.SSL{
-		Valid:     &valid,
-		ExpiresAt: &expiresAt,
-		Issuer:    &issuer,
-	}
-}
-
-func getCORSInfo(resp *http.Response) dto.CORS {
-	allowOrigin := resp.Header.Get("Access-Control-Allow-Origin")
-	enabled := allowOrigin != ""
-	if !enabled {
-		allowOrigin = "*"
-	}
-	return dto.CORS{
-		AllowOrigin: Ptr(allowOrigin),
-		Enabled:     Ptr(enabled),
-	}
-}
-
-func getDNSDetails(domain string) *dto.DnsRecords {
-	ips, err := net.LookupIP(domain)
-	if err != nil {
-		return nil
-	}
-
-	mxs, err := net.LookupMX(domain)
-	if err != nil {
-		return nil
-	}
-	mxDomains := make([]string, len(mxs))
-	for i, mx := range mxs {
-		mxDomains[i] = mx.Host
-	}
-
-	return &dto.DnsRecords{
-		A:     ips,
-		CNAME: Ptr(domain),
-		MX:    mxDomains,
-	}
-}
-
-func getServerInfo(domain string, resp *http.Response, dnsResponseTimeMs int64, dnsRecords *dto.DnsRecords) (*dto.ServerInfo, error) {
-	ipAddresses, err := net.LookupIP(domain)
-	if err != nil || len(ipAddresses) == 0 {
-		return nil, err
-	}
-
-	return &dto.ServerInfo{
-		IPAddress:         &ipAddresses[0],
-		WebServer:         getServerType(resp),
-		DnsResponseTimeMs: &dnsResponseTimeMs,
-		DnsRecords:        dnsRecords,
-	}, nil
-}
-
-func getServerType(resp *http.Response) *string {
-	server := resp.Header.Get("Server")
-	if server == "" {
-		return nil
-	}
-	return &server
-}
-
-func getOptimization(resp *http.Response) *string {
-	contentEncoding := resp.Header.Get("Content-Encoding")
-	if contentEncoding != "" {
-		return &contentEncoding
-	}
-
-	none := "none"
-	return &none
 }
